requirements.txt
streamlit
pandas
playwright
python-dotenv
openpyxl
